rfserver: Don't catch exceptions in main()
We have previously been catching exceptions in the main() function of
RFServer, and describing all failures as a failure to open the config
file. In the best case, this behaviour simply hides information that may
be useful for debugging. In the worst case, it misleads the user as to
where the problem is. This patch removes the code to catch these
exceptions, allowing the default handler to print the correct message.

Signed-off-by: Joe Stringer <joe@wand.net.nz>