Handle controller spawn and death
When a controller is spawned its details are registered with RFMonitor
and a `Monitor` is initialized for the controller that is responsible
for testing periodically if controller is alive. When a controller dies,
entries corresponding to the controller are removed from RFMonitor and
periodic test going on the controller is stopped.

Signed-off-by: Srijan Mishra <srijan23mishra@gmail.com>