rfclient: Rework port configuration messaging
Previously, when a virtual switch joined, the following occurred:-
1) VS notifies RFProxy
2) RFProxy notifies RFServer
3) RFServer notifies RFClient
4) RFClient sends a single ethernet packet out a port to determine VS
   port mapping
5) VS forwards to RFProxy with port information
6) RFProxy notifies RFServer
7) RFServer maps ports, tells RFProxy about VM-DP mapping
8) RFProxy stores VM-DP Mapping

This patch removes step 3, and makes step 7 also notify RFClient. When
RFClient is notified that the mapping has been successfully established,
it sets the interface state to "active". This allows cached routes with
egress via this interface to be pushed to the associated datapath.

Signed-off-by: Joe Stringer <joe@wand.net.nz>