rfclient: Refactor storage of Interface objects
This patch introduces a new abstract class, InterfaceMap and makes
RFClient subclass this. This allows the main thread to maintain the
collection of Interfaces and provide access to them in a thread-safe
manner. In future, this will assist with handling changes in Interface
status, and the ability to set ports administratively up/down.

FlowTable now makes use of this abstract class to get up-to-date
Interface information on-demand.

Signed-off-by: Joe Stringer <joe@wand.net.nz>