FlowTable: Keep track of deleted routes
If a route "add" occurs, and the route cannot be resolved, then
GWResolver will cache the route and continue attempting to insert it. If
the kernel attempts to delete this route while it is cached, then the
removal will be discarded, as the route does not exist in the routeTable.
However, the original "add" will remain cached, which causes a mismatch
between the kernel state and our final datapath state.

This patch adds a "deletedRoutes" map to keep track of these, and makes
GWResolver check this map for a given route before attempting to add it.
For route adds, this will remove the route from the deletedRoutes map
and discard the route. For route deletes, this will also remove the
route from the deletedRoutes map and attempt to remove the route from
the routeTable.

Signed-off-by: Joe Stringer <joe@wand.net.nz>