Master election by RFMonitor.
When a master controller dies a new master is to be elected among the
remaining controllers. A dictionary of eligible masters is maintained by
the RFMonitor which is updated periodically as a new controller comes or
the load on existing controllers increase. When a master dies a master
is selected from these eligible controllers. This patch implements the
required changes at RFMonitor level to elect a new master.

Signed-off-by: Srijan Mishra <srijan23mishra@gmail.com>