rftest1: Disable RFStats if using zeromq
RFStats relies on MongoDB to store state about mappings. If we're
using ZeroMQ, then this usually indicates that we are storing the state
in memory instead. In this case, the RFStats component will fail.

This patch checks if ZeroMQ is in use, and disables RFStats in this
case. This only affects pox-rfproxy, as RFStats is not currently ported
to other controllers.

Signed-off-by: Joe Stringer <joestringernz@gmail.com>