rfserver: better support for multiple vendors using multi-table datapaths
Until now, any multi-table datapath was assumed to be NoviFlow and other
vendors had to patch rfserver.py so that their custom RouteModTranslator
subclass would be used.

New format for --multitabledps option is:
  dpid1/vendor1,dpid2/vendor2,...

This allows the user to associate the vendor (and thus the appropriate
RouteModTranslator subclass) with a given datapath.